pyproject.toml
[project]
name = "gcp-release-notes"
version = "0.1.0"
description = "Streamlit app to search and filter Google Cloud release notes from BigQuery"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-bigquery>=3.31.0",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "plotly>=6.0.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.0",
    "streamlit-antd-components>=0.3.2",
]

[project.optional-dependencies]
vertex = [
    "google-cloud-aiplatform>=1.85.0",
    "vertexai>=1.43.0",
]

[dependency-groups]
dev = []
